refactor: moved hooks definition to types.ts
